fix(settings): make the column select a page, and give System its own
The environment configuration had three doors to one subject: a full-screen
route of its own, a link row inside settings, and the first-run gate. It now
has one. It is a settings page like Account or Appearance, opened the way the
others are.

That required the column to actually be navigation. It scrolled to a heading
while an IntersectionObserver guessed which heading was in view to light the
right entry, so every section was mounted at once in a single page that grows
without bound as settings are added. One page renders at a time now, which is
what the column already looked like it did, and the observer, the scroll and
the refs that fed them are gone.

Within System the panel is the page body rather than the value of a labelled
row: wrapped in a SettingRow it printed its title twice and indented the whole
thing as though it were one control. Sections no longer carry their own
heading, since the body titles the page once.

The gate opens settings at System instead of a screen with no way back to
anything else, via goProfile(page). AppScreen loses "environment", and with it
goEnvironment and the title bar eyebrow for a screen that no longer exists.

Verified: tsc, vite build, go build, go test, 193 sidecar tests.
