+++ frontend/src/components/EnvironmentPanel.tsx
@@ -1,0 +1,394 @@
+/**
+ * The Python environment, as a section of the settings screen.
+ *
+ * A section rather than a destination of its own: choosing which interpreter
+ * runs the analyses is a setting, and it belongs beside the other settings.
+ * Given its own place in the navigation it became a third door to one thing,
+ * next to the settings row and the first-run gate.
+ *
+ * This is what the LITE install never had. Its path asked a user to install
+ * Python, make a virtual environment, install a requirements file that is not
+ * in the download, and export a variable that a desktop launch never reads --
+ * and then, if any of that went wrong, said nothing until an analysis died
+ * mid-run on an import.
+ *
+ * So it states what is true rather than what is convenient: which interpreter
+ * is in use and how it was picked, which dependency is missing and what that
+ * specifically stops working, and whether an environment variable is
+ * overriding the choice offered here.
+ */
+import { useCallback, useEffect, useRef, useState } from "react"
+import { AlertTriangle, Check, Loader2, RefreshCw, X } from "lucide-react"
+import { EventsOn } from "../../wailsjs/runtime/runtime"
+import {
+  BuildManagedEnvironment,
+  CancelEnvironmentBuild,
+  InspectEnvironment,
+  UseInterpreter,
+} from "../../wailsjs/go/main/App"
+import type { backend, main } from "../../wailsjs/go/models"
+import { btnGhost, btnPrimaryCommit } from "@/components/ui/buttons"
+import { cn } from "@/lib/utils"
+
+type SetupEvent = {
+  step: string
+  line?: string
+  error?: string
+}
+
+/** How an interpreter came to be the one in use, in the user's terms. */
+const ORIGIN_LABEL: Record<string, string> = {
+  managed: "created by TERRA",
+  bundled: "shipped with TERRA",
+  chosen: "chosen here",
+  venv: "a project environment",
+  path: "found on this machine",
+  detected: "detected automatically",
+  GEOSENSE_PYTHON: "forced by GEOSENSE_PYTHON",
+}
+
+/*
+  Takes no callback for "the environment now works".
+
+  It had one while this was a full-screen route that the first-run gate opened
+  and had to be dismissed from. As a settings page there is nothing to dismiss:
+  the user is somewhere they can stay, and the panel reporting its own new state
+  is the whole of the feedback.
+*/
+export function EnvironmentPanel() {
+  const [state, setState] = useState<main.EnvironmentState | null>(null)
+  const [loading, setLoading] = useState(true)
+  const [busyPath, setBusyPath] = useState<string | null>(null)
+  const [problem, setProblem] = useState<string | null>(null)
+  const [log, setLog] = useState<SetupEvent[]>([])
+  const [building, setBuilding] = useState(false)
+  const logRef = useRef<HTMLDivElement | null>(null)
+
+  const refresh = useCallback(async () => {
+    setLoading(true)
+    try {
+      setState(await InspectEnvironment())
+      setProblem(null)
+    } catch (e) {
+      setProblem(e instanceof Error ? e.message : String(e))
+    } finally {
+      setLoading(false)
+    }
+  }, [])
+
+  useEffect(() => {
+    void refresh()
+  }, [refresh])
+
+  // pip's own lines, as they arrive. An install takes minutes, and its output
+  // is the only honest answer to "is this stuck".
+  useEffect(() => {
+    return EventsOn("env:setup", (ev: SetupEvent) => {
+      setLog((prev) => [...prev.slice(-400), ev])
+      if (ev.step === "failed") {
+        setBuilding(false)
+        setProblem(ev.error ?? "the environment could not be built")
+      } else if (ev.step === "done") {
+        setBuilding(false)
+        void refresh()
+      }
+    })
+  }, [refresh])
+
+  useEffect(() => {
+    logRef.current?.scrollTo({ top: logRef.current.scrollHeight })
+  }, [log])
+
+  const build = async (basePython: string) => {
+    setLog([])
+    setProblem(null)
+    setBuilding(true)
+    try {
+      await BuildManagedEnvironment(basePython)
+    } catch (e) {
+      setBuilding(false)
+      setProblem(e instanceof Error ? e.message : String(e))
+    }
+  }
+
+  const choose = async (path: string) => {
+    setBusyPath(path)
+    setProblem(null)
+    try {
+      await UseInterpreter(path)
+      await refresh()
+    } catch (e) {
+      setProblem(e instanceof Error ? e.message : String(e))
+    } finally {
+      setBusyPath(null)
+    }
+  }
+
+  const active = state?.active ?? null
+  const missing = (active?.packages ?? []).filter(
+    (p) => !p.optional && (!p.present || p.version_problem)
+  )
+  const degraded = (active?.packages ?? []).filter((p) => p.optional && !p.present)
+
+  return (
+    <div className="flex flex-col gap-3">
+      {state?.env_override && (
+        <Notice tone="warning">
+          <span className="telemetry">GEOSENSE_PYTHON</span> is set to{" "}
+          <span className="telemetry">{state.env_override}</span>, and it overrides
+          everything chosen here. Unset it to let this screen decide.
+        </Notice>
+      )}
+      {problem && <Notice tone="error">{problem}</Notice>}
+
+      {/* What is in use, and whether it works. */}
+      <section className="rounded-sm border border-border bg-secondary/50 p-4">
+        <p className="eyebrow mb-3">In use</p>
+        {loading && !active ? (
+          <p className="flex items-center gap-2 text-body text-muted-foreground">
+            <Loader2 className="size-3.5 animate-spin" />
+            checking
+          </p>
+        ) : !active ? (
+          <p className="text-body text-muted-foreground">
+            No interpreter is resolved yet.
+          </p>
+        ) : (
+          <>
+            <div className="flex flex-wrap items-baseline justify-between gap-2">
+              <span className="telemetry min-w-0 truncate text-emphasis text-foreground">
+                {active.executable}
+              </span>
+              <StatusPill
+                ok={active.usable}
+                label={
+                  active.unreachable
+                    ? "cannot be run"
+                    : active.usable
+                      ? `ready · python ${active.python_version}`
+                      : `python ${active.python_version} · incomplete`
+                }
+              />
+            </div>
+            <p className="mt-1 text-body text-muted-foreground">
+              {ORIGIN_LABEL[active.origin] ?? active.origin}
+              {!active.python_ok &&
+                ` · TERRA needs Python ${active.min_python} or newer`}
+            </p>
+
+            {missing.length > 0 && (
+              <ul className="mt-3 flex flex-col gap-1.5">
+                {missing.map((p) => (
+                  <PackageRow key={p.module} pkg={p} />
+                ))}
+              </ul>
+            )}
+            {degraded.length > 0 && (
+              <ul className="mt-3 flex flex-col gap-1.5">
+                {degraded.map((p) => (
+                  <PackageRow key={p.module} pkg={p} />
+                ))}
+              </ul>
+            )}
+          </>
+        )}
+      </section>
+
+      {/* The recommended repair, and the interpreters it can be built on. */}
+      <section className="rounded-sm border border-border bg-secondary/50 p-4">
+        <p className="eyebrow mb-1">Interpreters on this machine</p>
+        <p className="mb-3 text-body text-muted-foreground">
+          TERRA can build its own environment from any Python{" "}
+          {active?.min_python ?? "3.12"} or newer, and keep it beside its database
+          rather than inside the application — so an update does not discard it.
+        </p>
+
+        <ul className="flex flex-col gap-2">
+          {(state?.candidates ?? []).map((c) => (
+            <CandidateRow
+              key={c.path}
+              candidate={c}
+              active={c.path === active?.executable}
+              disabled={building || busyPath !== null}
+              busy={busyPath === c.path}
+              onBuild={() => void build(c.path)}
+              onChoose={() => void choose(c.path)}
+            />
+          ))}
+          {(state?.candidates ?? []).length === 0 && !loading && (
+            <li className="text-body text-muted-foreground">
+              No Python was found. Install Python {active?.min_python ?? "3.12"} and
+              check again.
+            </li>
+          )}
+        </ul>
+
+        <div className="mt-3 flex items-center gap-2">
+          <button
+            type="button"
+            onClick={() => void refresh()}
+            disabled={loading || building}
+            className={btnGhost}
+          >
+            <RefreshCw className={cn("size-3.5", loading && "animate-spin")} />
+            Check again
+          </button>
+          {building && (
+            <button
+              type="button"
+              onClick={() => void CancelEnvironmentBuild()}
+              className={btnGhost}
+            >
+              <X className="size-3.5" />
+              Stop
+            </button>
+          )}
+        </div>
+      </section>
+
+      {log.length > 0 && (
+        <section className="rounded-sm border border-border bg-secondary/50 p-4">
+          <p className="eyebrow mb-3">{building ? "Installing" : "Last install"}</p>
+          {/* Verbatim, not summarised. What pip is resolving is the only
+                  honest answer to why this is taking minutes, and on failure
+                  the reason is in these lines. */}
+          <div
+            ref={logRef}
+            className="panel-scroll max-h-56 overflow-y-auto rounded-sm border border-border bg-background p-2"
+          >
+            {log.map((ev, i) => (
+              <p
+                key={i}
+                className={cn(
+                  "telemetry whitespace-pre-wrap break-all text-micro leading-relaxed",
+                  ev.error ? "text-destructive-quiet" : "text-muted-foreground"
+                )}
+              >
+                {ev.error ?? ev.line}
+              </p>
+            ))}
+          </div>
+        </section>
+      )}
+    </div>
+  )
+}
+
+function StatusPill({ ok, label }: { ok: boolean; label: string }) {
+  return (
+    <span
+      className={cn(
+        "telemetry flex shrink-0 items-center gap-1.5 rounded-sm px-1.5 py-0.5 text-meta",
+        ok
+          ? "bg-primary/20 text-accent-quiet"
+          : "bg-secondary text-muted-foreground"
+      )}
+    >
+      {ok ? <Check className="size-3" /> : <AlertTriangle className="size-3" />}
+      {label}
+    </span>
+  )
+}
+
+/**
+ * One dependency, named by what its absence costs.
+ *
+ * "pvlib is missing" means nothing to an agronomist. "the photovoltaic model"
+ * is the thing they came for, and the sidecar states it, so it is what shows.
+ */
+function PackageRow({ pkg }: { pkg: backend.EnvPackage }) {
+  return (
+    <li className="flex flex-wrap items-baseline gap-x-2 gap-y-0.5 text-body">
+      <span
+        className={cn(
+          "telemetry shrink-0 rounded-sm px-1 text-micro uppercase",
+          pkg.optional
+            ? "border border-border text-muted-foreground"
+            : "bg-destructive-quiet/15 text-destructive-quiet"
+        )}
+      >
+        {pkg.optional ? "degraded" : "missing"}
+      </span>
+      <span className="telemetry text-foreground">{pkg.distribution}</span>
+      {pkg.version_problem ? (
+        <span className="text-muted-foreground">
+          {pkg.version_problem}
+          {pkg.wanted && <> · needs {pkg.wanted}</>}
+          {pkg.why && <> · {pkg.why}</>}
+        </span>
+      ) : (
+        <span className="text-muted-foreground">without it: {pkg.blocks}</span>
+      )}
+    </li>
+  )
+}
+
+function CandidateRow({
+  candidate,
+  active,
+  disabled,
+  busy,
+  onBuild,
+  onChoose,
+}: {
+  candidate: backend.PythonCandidate
+  active: boolean
+  disabled: boolean
+  busy: boolean
+  onBuild: () => void
+  onChoose: () => void
+}) {
+  return (
+    <li className="flex flex-wrap items-center justify-between gap-2 rounded-sm border border-border bg-background px-3 py-2">
+      <div className="min-w-0">
+        <p className="telemetry truncate text-body text-foreground">
+          {candidate.path}
+        </p>
+        <p className="text-micro text-muted-foreground">
+          {ORIGIN_LABEL[candidate.origin] ?? candidate.origin}
+          {active && " · in use"}
+        </p>
+      </div>
+      <div className="flex shrink-0 items-center gap-2">
+        <button
+          type="button"
+          onClick={onChoose}
+          disabled={disabled || active}
+          className={btnGhost}
+        >
+          {busy ? <Loader2 className="size-3.5 animate-spin" /> : null}
+          Use as is
+        </button>
+        <button
+          type="button"
+          onClick={onBuild}
+          disabled={disabled}
+          className={btnPrimaryCommit}
+        >
+          Build environment
+        </button>
+      </div>
+    </li>
+  )
+}
+
+function Notice({
+  tone,
+  children,
+}: {
+  tone: "warning" | "error"
+  children: React.ReactNode
+}) {
+  return (
+    <p
+      className={cn(
+        "rounded-sm border px-3 py-2 text-body",
+        tone === "error"
+          ? "border-destructive-quiet/40 text-destructive-quiet"
+          : "border-border text-muted-foreground"
+      )}
+    >
+      {children}
+    </p>
+  )
+}
