+++ frontend/src/components/TitleBar.tsx
@@ -43,7 +43,6 @@ const SCREEN_EYEBROW: Record<AppScreen, string> = {
   analysis: "project hub",
   auth: "sign in",
   profile: "settings",
-  environment: "python environment",
 }
 
 function fmtCoord(v: number, pos: string, neg: string): string {
